-finished all controllers and services
-fix delete cascade in database
-added interfaces to services
-changed money fields types to decimal
-deleted all empty lines
-updated response models

+++ HotelReservation/Api/Models/RequestModels/RoomRequestModel.cs
@@ -9,6 +9,6 @@ namespace HotelReservation.Api.Models.RequestModels
     {
         public string RoomNumber { get; set; }
         public int BedsNumber { get; set; }
-        public double PaymentPerDay { get; set; }
+        public decimal PaymentPerDay { get; set; }
     }
 }
