+++ HotelReservation/Api/Models/RequestModels/ServiceRequestModel.cs
@@ -8,6 +8,6 @@ namespace HotelReservation.Api.Models.RequestModels
     public class ServiceRequestModel
     {
         public string Name { get; set; }
-        public double Payment { get; set; }
+        public decimal Payment { get; set; }
     }
 }
