+++ HotelReservation/Api/Models/ResponseModels/OrderResponseModel.cs
@@ -14,6 +14,6 @@ namespace HotelReservation.Api.Models.ResponseModels
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
         public int NumberOfDays { get; set; }
-        public double FullPrice { get; set; }
+        public decimal FullPrice { get; set; }
     }
 }
