+++ HotelReservation/Data/Entities/OrderEntity.cs
@@ -16,7 +16,7 @@ namespace HotelReservation.Data.Entities
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
         public int NumberOfDays { get; set; }
-        public double FullPrice { get; set; }
+        public decimal FullPrice { get; set; }
 
     }
 }
